+++ src/Popups/Credits.js
@@ -95,6 +95,33 @@ function CreditsPopup(props) {
       <DialogContent dividers className={classes.dialog}>
         <div hidden={expanded !== 0}>
           <Paper className={classes.content}>
+            <Typography variant="h5" className={classes.version}>v1.7.0 (2021-10-09)</Typography>
+            <Typography variant="h6">New</Typography>
+            <List dense>
+              <ListItem>New custom layers: add custom layers to the map with your own filters and display settings</ListItem>
+              <ListItem>New basemap with English location names</ListItem>
+              <ListItem>New ILS filter: only display and use airports that have an ILS approach (MSFS). Thanks to Lily418 for the help</ListItem>
+              <ListItem>New setting in Route Finder to set a custom airplane rental price (<IssueLink id={65} />)</ListItem>
+            </List>
+            <Typography variant="h6">Changed</Typography>
+            <List dense>
+              <ListItem>Optimized Route Finder memory usage: this should prevent crashes with Chrome, even when searching large areas</ListItem>
+              <ListItem>Improved Route Finder when setting a destination: no more route going in the wrong overall direction</ListItem>
+              <ListItem>Optimized application memory usage and loading time</ListItem>
+              <ListItem>Updated runway data (length and surface)</ListItem>
+              <ListItem>Updated plane list to include newly added FSE planes</ListItem>
+              <ListItem>Updated MSFS data</ListItem>
+              <ListItem>Changed display in Route Finder PDF to separate cargo weight from the total weight (cargo and passengers) (<IssueLink id={66} />)</ListItem>
+            </List>
+            <Typography variant="h6">Fixed</Typography>
+            <List dense>
+              <ListItem>Bug when loading data from airports with an ICAO resembling a number, such as 0E0 (<IssueLink id={79} />)</ListItem>
+              <ListItem>Rounded airplane specs in Route Finder for a better display</ListItem>
+              <ListItem>Bug when resetting settings multiple times</ListItem>
+            </List>
+          </Paper>
+
+          <Paper className={classes.content}>
             <Typography variant="h5" className={classes.version}>v1.6.0 (2021-06-27)</Typography>
             <Typography variant="h6">New</Typography>
             <List dense>
@@ -416,6 +443,8 @@ function CreditsPopup(props) {
             <Typography variant="h5" className={classes.version}>Contributors</Typography>
             <List dense>
               <ListItem>piero-la-lune (author)</ListItem>
+              <ListItem>icykoneko (contributor)</ListItem>
+              <ListItem>Lily418 (author)</ListItem>
             </List>
           </Paper>